Tpetra: Second chunk of triple product code refactor
